calibrate/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "calibrate-pro"
version = "1.0.0"
description = "Professional sensorless display calibration"
readme = "README.md"
license = {text = "Proprietary"}
authors = [
    { name = "Zain Dana Quanta" },
]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.optional-dependencies]
gui = ["PyQt6>=6.0"]
tray = ["pystray>=0.19", "Pillow>=9.0"]
sensor = ["hidapi>=0.14"]
macos = ["pyobjc-framework-Quartz>=9.0", "pyobjc-framework-CoreFoundation>=9.0"]
all = ["PyQt6>=6.0", "pystray>=0.19", "Pillow>=9.0", "hidapi>=0.14"]

[project.scripts]
calibrate-pro = "calibrate_pro.main:main"

[tool.setuptools.packages.find]
include = ["calibrate_pro*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
